pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "provisioner"
version = "0.1.0"
description = "ACC/FORMA user management CLI (APS)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
